Optimize inference loop and compilation paths for AMD ROCm
- Eliminate CPU-GPU synchronization bottlenecks in the generate loop by
  vectorizing the repetition penalty (.unique()) and removing .item()
  syncs during EOS checks.
- Prevent dynamic shape compilation errors under PyTorch Inductor by
  restricting compile to the decode loop and removing data-dependent
  guards (T > 1) in the attention layer.
- Swap out-of-the-box RMSNorm with PyTorch's native F.rms_norm fused
  kernel to minimize memory bandwidth on APUs.
- Prune Dropout(p=0.0) layers with Identity before compilation to
  allow for cleaner Triton kernel fusion.
- Default Gradio UI precision to float16 and initialize TunableOp
  environment variables before PyTorch imports.

+++ model_runner/app.py
@@ -508,7 +508,7 @@ def create_ui():
                 precision_dropdown = gr.Dropdown(
                     label="Precision Mode",
                     choices=["float32", "bfloat16", "float16", "int8", "int4"],
-                    value="float32",
+                    value="float16",
                     interactive=True,
                     scale=2,
                 )
